Configurable factories.
