+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/PersistentStorageFactory.java
@@ -16,67 +16,15 @@
 
 package com.palantir.atlasdb.factory;
 
-import java.io.File;
-import java.util.UUID;
-import java.util.regex.Pattern;
-
-import org.rocksdb.RocksDB;
-import org.rocksdb.RocksDBException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.google.common.base.MoreObjects;
 import com.palantir.atlasdb.config.RocksDbPersistentStorageConfig;
 import com.palantir.atlasdb.persistent.api.PersistentTimestampStore;
-import com.palantir.atlasdb.persistent.rocksdb.RocksDbPersistentTimestampStore;
-import com.palantir.logsafe.Preconditions;
 
-public final class PersistentStorageFactory {
-    private static final Logger log = LoggerFactory.getLogger(PersistentStorageFactory.class);
-    private static final Pattern UUID_PATTERN = Pattern.compile(
-            "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$");
-
+public interface PersistentStorageFactory {
     /**
-     * For the given path does the following: 1) if it exists checks that it is a directory, 2) if it is a directory
-     * removes all sub-folders whose names are string representation of a UUID.
-     *
-     * @param storagePath to the proposed storage location
+     * Constructs a {@link com.palantir.atlasdb.persistent.rocksdb.RocksDbPersistentTimestampStore} using the
+     * supplied configuration.
+     * @param config to use to configure the store
+     * @return store to be used
      */
-    public static void sanitizeStoragePath(String storagePath) {
-        File storageDirectory = new File(storagePath);
-        if (!storageDirectory.exists()) {
-            Preconditions.checkState(storageDirectory.mkdir(), "Not able to create a storage directory");
-            return;
-        }
-        Preconditions.checkArgument(storageDirectory.isDirectory(), "Storage path has to be a directory");
-        for (File file : MoreObjects.firstNonNull(storageDirectory.listFiles(), new File[0])) {
-            if (file.isDirectory()) {
-                if (UUID_PATTERN.matcher(file.getName()).matches()) {
-                    file.delete();
-                }
-            }
-        }
-    }
-
-    /**
-     * Constructs a {@link PersistentTimestampStore} from a {@link RocksDbPersistentStorageConfig}.
-     *
-     * @param config of the requested RocksDB persistent storage
-     * @return RockDB implementation of {@link PersistentTimestampStore}
-     */
-    public PersistentTimestampStore constructPersistentTimestampStore(RocksDbPersistentStorageConfig config) {
-        File databaseFolder = new File(config.storagePath(), UUID.randomUUID().toString());
-        Preconditions.checkState(databaseFolder.mkdir(), "RocksDb directory created");
-        RocksDB rocksDb = openRocksConnection(databaseFolder);
-        return new RocksDbPersistentTimestampStore(rocksDb, databaseFolder);
-    }
-
-    private static RocksDB openRocksConnection(File databaseFolder) {
-        try {
-            return RocksDB.open(databaseFolder.getAbsolutePath());
-        } catch (RocksDBException e) {
-            log.error("Opening RocksDB failed", e);
-            throw new RuntimeException(e);
-        }
-    }
+    PersistentTimestampStore constructPersistentTimestampStore(RocksDbPersistentStorageConfig config);
 }
